refactor(core): extract feed pipeline from Html2rss entrypoint
